Moved checkIcon function across to new image module, removed old images module.
