--- src/_shared/js/images.js
@@ -1,9 +1,0 @@
-export function checkIcon(responseJson) {
-    return responseJson.audioTag ?
-        audioIcon :
-    responseJson.galleryTag ?
-        imageIcon :
-    responseJson.videoTag ?
-        videoIcon :
-        '';
-}
