+++ src/capi-single-paid/index.js
@@ -2,8 +2,7 @@ import { enableToggles } from '../_shared/js/ui';
 import { write } from '../_shared/js/dom';
 import { getIframeId, getWebfonts, resizeIframeHeight } from '../_shared/js/messages';
 import { getApiBaseUrl } from '../_shared/js/dev';
-import { checkIcon } from '../_shared/js/images';
-import { insertImage } from '../_shared/js/capi-images.js';
+import { insertImage, checkIcon } from '../_shared/js/capi-images.js';
 
 let container = document.getElementsByClassName('adverts__body')[0];
 let params = new URLSearchParams();
