+++ src/capi-single-supported/index.js
@@ -1,8 +1,7 @@
 import { enableToggles } from '../_shared/js/ui';
 import { write } from '../_shared/js/dom';
 import { getIframeId, getWebfonts, resizeIframeHeight } from '../_shared/js/messages';
-import { checkIcon } from '../_shared/js/images';
-import { insertImage } from '../_shared/js/capi-images.js';
+import { insertImage, checkIcon } from '../_shared/js/capi-images.js';
 
 let container = document.getElementsByClassName('adverts--supported')[0];
 let params = new URLSearchParams();
